+++ src/main/java/com/gw/forum/forum/controller/PublishController.java
@@ -35,7 +35,7 @@ public class PublishController {
     public  String doPublish(@RequestParam(name = "title",required = false)String title,
                              @RequestParam(name = "description",required = false)String description,
                              @RequestParam(name = "tag",required = false)String tag,
-                             @RequestParam(name = "id",required = false)Integer id,
+                             @RequestParam(name = "id",required = false) Long id,
                              HttpServletRequest request,
                              Model model){
         model.addAttribute("title",title);
@@ -73,7 +73,7 @@ public class PublishController {
         return "publish";
     }
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable("id")Integer id,
+    public String edit(@PathVariable("id") Long id,
                        Model model){
         QuestionDTO question= questionService.revertPage(id);
         model.addAttribute("title",question.getTitle());
