+++ src/main/java/com/gw/forum/forum/controller/RevertController.java
@@ -13,7 +13,7 @@ public class RevertController {
     @Autowired
     private QuestionService questionService;
     @GetMapping("/revert/{id}")
-    public String revert(@PathVariable("id")Integer id,
+    public String revert(@PathVariable("id") Long id,
                          Model model){
         QuestionDTO questionDTO=questionService.revertPage(id);
         questionService.incView(id);
